+++ packages/expo-router/src/link/__tests__/Link.test.ios.tsx
@@ -343,7 +343,6 @@ describe('singular', () => {
     expect(screen).toHaveRouterState({
       index: 0,
       key: expect.any(String),
-      preloadedRoutes: [],
       routeNames: ['__root', '+not-found', '_sitemap'],
       routes: [
         {
@@ -355,7 +354,6 @@ describe('singular', () => {
           state: {
             index: 3,
             key: expect.any(String),
-            preloadedRoutes: [],
             routeNames: ['[slug]'],
             routes: [
               {
@@ -406,7 +404,6 @@ describe('singular', () => {
     expect(screen).toHaveRouterState({
       index: 0,
       key: expect.any(String),
-      preloadedRoutes: [],
       routeNames: ['__root', '+not-found', '_sitemap'],
       routes: [
         {
@@ -418,7 +415,6 @@ describe('singular', () => {
           state: {
             index: 1,
             key: expect.any(String),
-            preloadedRoutes: [],
             routeNames: ['[slug]'],
             routes: [
               {
@@ -474,7 +470,6 @@ test('can dynamically route using singular function', () => {
   expect(screen).toHaveRouterState({
     index: 0,
     key: expect.any(String),
-    preloadedRoutes: [],
     routeNames: ['__root', '+not-found', '_sitemap'],
     routes: [
       {
@@ -486,7 +481,6 @@ test('can dynamically route using singular function', () => {
         state: {
           index: 4,
           key: expect.any(String),
-          preloadedRoutes: [],
           routeNames: ['[slug]'],
           routes: [
             {
@@ -548,7 +542,6 @@ test('can dynamically route using singular function', () => {
   expect(screen).toHaveRouterState({
     index: 0,
     key: expect.any(String),
-    preloadedRoutes: [],
     routeNames: ['__root', '+not-found', '_sitemap'],
     routes: [
       {
@@ -560,7 +553,6 @@ test('can dynamically route using singular function', () => {
         state: {
           index: 3,
           key: expect.any(String),
-          preloadedRoutes: [],
           routeNames: ['[slug]'],
           routes: [
             {
@@ -620,7 +612,6 @@ describe('prefetch', () => {
     expect(screen).toHaveRouterState({
       index: 0,
       key: expect.any(String),
-      preloadedRoutes: [],
       routeNames: ['__root', '+not-found', '_sitemap'],
       routes: [
         {
@@ -630,13 +621,6 @@ describe('prefetch', () => {
           state: {
             index: 0,
             key: expect.any(String),
-            preloadedRoutes: [
-              {
-                key: expect.any(String),
-                name: 'test',
-                params: {},
-              },
-            ],
             routeNames: ['index', 'test'],
             routes: [
               {
@@ -644,6 +628,11 @@ describe('prefetch', () => {
                 name: 'index',
                 params: undefined,
                 path: '/',
+              },
+              {
+                key: expect.any(String),
+                name: 'test',
+                params: {},
               },
             ],
             stale: false,
@@ -680,7 +669,13 @@ describe('prefetch', () => {
     if (innerState?.type !== 'stack') {
       throw new Error('Expected a stack navigator');
     }
-    expect((innerState as StackNavigationState<ParamListBase>).preloadedRoutes).toEqual([
+    expect((innerState as StackNavigationState<ParamListBase>).routes).toEqual([
+      {
+        key: expect.stringMatching(/^index-/),
+        name: 'index',
+        params: undefined,
+        path: '/',
+      },
       {
         key: expect.stringMatching(/^test-/),
         name: 'test',