+++ packages/expo-router/src/react-navigation/routers/__tests__/getRoutesForRouteNames.test.tsx
@@ -15,7 +15,6 @@ function stackState(
     routes,
     type: 'stack',
     stale: false,
-    preloadedRoutes: [],
   } as StackNavigationState<ParamListBase>;
 }